fix(auth) redis에 RT 삭제되지 않는 오류 수정 및 토큰 관련 주석 수정

+++ auth-service/src/main/java/com/green/auth/application/auth/AuthMemberRepository.java
@@ -3,6 +3,6 @@ package com.green.auth.application.auth;
 import com.green.auth.entity.AuthMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthMemberRepository extends JpaRepository<AuthMember, Integer> {
+public interface AuthMemberRepository extends JpaRepository<AuthMember, Long> {
     boolean existsByMemberCodeAndEmail(Integer memberCode, String email);
 }
