+++ auth-service/src/main/java/com/green/auth/application/auth/model/AuthMemberCreateReq.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class AuthMemberCreateReq {
-    private Integer memberCode;
+    private Long memberCode;
     private String password;
     private EnumMemberRole role;
     private String email;
