+++ auth-service/src/main/java/com/green/auth/application/auth/model/LoginReq.java
@@ -4,6 +4,6 @@ import lombok.Getter;
 
 @Getter
 public class LoginReq {
-    private Integer memberCode;
+    private Long memberCode;
     private String password;
 }
